Finished with AddPart Form/Controller and ModifyPart Form/Controller. The following changes:
HUGE FIX: Was having problems with starting the application. TableView would not load test data. The Fix was updating the module to include the model to javax.fxml and exporting to model as well. Seems like I forgot to add this to the module in the beginning. Hence, the module could not load the test data into the tableview.

For AddPart Controller:
- Fixed the error popping everytime when saving
- Added on self-incrementing ID for the product adding
- Finished Documentation

For ModifyPart Controller:
- Implemented Setter for Data object passing (selected item from the Main Form)
- Finished coding the object addition to the Inventory based off object attributes
- Finished Documentation with code and also mentioned logical errors within codes that was fixed.
- Added on logical error fixes with user inputs and also integrated self test for each data entry,

Next Section to work on: Add Product / Modify Product.
Things to look out for: Association with Products and Parts.